requirements.txt
streamlit 
langchain_groq
langchain_community
python-dotenv
